+++ JewelOfIndiaBuilder/Controllers/PropertyAdminController.cs
@@ -14,9 +14,15 @@ namespace JewelOfIndiaBuilder.Controllers
     {
         private JewelOfIndiaEntities db = new JewelOfIndiaEntities();
 
+        public bool Redirect()
+        {
+            return !ApplicationSecurity.CheckUser(Session["UserType"].ToString(), Session["IsAdmin"].ToString(), "PROPERTY");
+        }
         // GET: /PropertyAdmin/
         public ActionResult Index()
         {
+            if (Redirect())
+                return RedirectToAction("../Home");
             var properties = db.Properties.Include(p => p.Location).Include(p => p.PropertyType);
             return View(properties.ToList());
         }
@@ -24,6 +30,8 @@ namespace JewelOfIndiaBuilder.Controllers
         // GET: /PropertyAdmin/Details/5
         public ActionResult Details(long? id)
         {
+            if (Redirect())
+                return RedirectToAction("../Home");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,6 +47,8 @@ namespace JewelOfIndiaBuilder.Controllers
         // GET: /PropertyAdmin/Create
         public ActionResult Create()
         {
+            if (Redirect())
+                return RedirectToAction("../Home");
             ViewBag.LocationId = new SelectList(db.Locations, "Id", "Address");
             ViewBag.PropertyTypeId = new SelectList(db.PropertyTypes, "Id", "CodeReference");
             return View();
@@ -51,6 +61,8 @@ namespace JewelOfIndiaBuilder.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include="Id,Feature,Description,LocationId,PropertyTypeId")] Property property)
         {
+            if (Redirect())
+                return RedirectToAction("../Home");
             if (ModelState.IsValid)
             {
                 db.Properties.Add(property);
@@ -66,6 +78,8 @@ namespace JewelOfIndiaBuilder.Controllers
         // GET: /PropertyAdmin/Edit/5
         public ActionResult Edit(long? id)
         {
+            if (Redirect())
+                return RedirectToAction("../Home");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -87,6 +101,8 @@ namespace JewelOfIndiaBuilder.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include="Id,Feature,Description,LocationId,PropertyTypeId")] Property property)
         {
+            if (Redirect())
+                return RedirectToAction("../Home");
             if (ModelState.IsValid)
             {
                 db.Entry(property).State = System.Data.Entity.EntityState.Modified;
@@ -101,6 +117,8 @@ namespace JewelOfIndiaBuilder.Controllers
         // GET: /PropertyAdmin/Delete/5
         public ActionResult Delete(long? id)
         {
+            if (Redirect())
+                return RedirectToAction("../Home");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -118,6 +136,8 @@ namespace JewelOfIndiaBuilder.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(long id)
         {
+            if (Redirect())
+                return RedirectToAction("../Home");
             Property property = db.Properties.Find(id);
             db.Properties.Remove(property);
             db.SaveChanges();
